FIX improved dropdown positioning in DataSpreadSheet and DataImporter
